refactor(admin): colocate live logs view and stream state module
The log viewer page and the websocket-backed logStream state lived in view/ and state/. Move both under admin/logs so log streaming, error badges and the viewer page are maintained together.
