+++ webui/src/admin/logs/Logs.vue
@@ -8,8 +8,8 @@
       <div ref="bodyEl" class="task-terminal-body log-page-body">
         <div v-if="logs.length === 0" class="task-terminal-hint">等待日志输出…</div>
         <div
-          v-for="(entry, index) in logs"
-          :key="`${entry.timestamp}-${index}`"
+          v-for="entry in logs"
+          :key="entry.seq"
           class="task-terminal-line"
           :class="logLevelClass(entry.level)"
         >
@@ -26,7 +26,7 @@
 import { nextTick, onMounted, onUnmounted, ref, watch } from "vue";
 
 import AdminPageHeader from "../components/AdminPageHeader.vue";
-import { clearLogs, enterLogsPage, leaveLogsPage, logLevelClass, logs } from "../state/logStream";
+import { clearLogs, enterLogsPage, leaveLogsPage, logLevelClass, logs } from "./logStream";
 
 const bodyEl = ref<HTMLElement | null>(null);
 
@@ -34,11 +34,22 @@ function scrollToBottom(): void {
   if (bodyEl.value) bodyEl.value.scrollTop = bodyEl.value.scrollHeight;
 }
 
+// 距底部 30px 内视为贴底：给平滑滚动动画和像素取整留容差
+function isNearBottom(): boolean {
+  const el = bodyEl.value;
+  if (!el) return true;
+  return el.scrollTop + el.clientHeight >= el.scrollHeight - 30;
+}
+
+// 追踪最后一条的 seq 而非数组长度：日志达到 500 条上限后 push+裁剪使长度不变，
+// 监听 length 会在稳定状态下漏掉新日志
 watch(
-  () => logs.value.length,
+  () => logs.value[logs.value.length - 1]?.seq,
   async () => {
+    // watch 默认 flush: 'pre'，回调时 DOM 尚未更新，读到的是新日志渲染前用户所处的位置
+    const followBottom = isNearBottom();
     await nextTick();
-    scrollToBottom();
+    if (followBottom) scrollToBottom();
   }
 );
 
@@ -54,7 +65,7 @@ onUnmounted(() => {
 </script>
 
 <style scoped lang="scss">
-@use "../styles/tasks" as *;
+@use "../tasks/tasks" as *;
 
 .log-page-body {
   height: min(82vh, 920px);
